% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.iDRWP.R
\name{get.iDRWP}
\alias{get.iDRWP}
\title{get iDRW-generated pathway profile}
\usage{
get.iDRWP(x, y, globalGraph, pathSet, class.outcome, covs = NULL,
  family = "cox", Gamma = 0.3, Corr = FALSE)
}
\arguments{
\item{x}{list of feature matrices (genomic profiles). The length of \code{x} >= 1. Each \code{x} should be samples as rows and genes (features) as columns.}

\item{y}{clinical matrix that contains response variable. The type of response variable can be either survival time or binary/multinomial response.}

\item{globalGraph}{directed multi-layered gene-gene graph.}

\item{pathSet}{The set of pathways.}

\item{class.outcome}{The variable name of response variable. For \code{class.outcome} is survival time \code{time}, two columns of survival time (\code{time}) and event status (\code{status}) should be provided.}

\item{covs}{A vector of variable names of covariates (optional).}

\item{family}{The type of response variable - \code{cox} (default) / \code{binomial} / \code{multinomial}.}

\item{Gamma}{The restart probability of random walks (default: \code{0.3}).}

\item{Corr}{A logical variable to indicate between-layer interactions are defined by correlation between genes (default: \code{FALSE}).}
}
\value{
A list of following elements.
\itemize{
\item pathActivity : A pathway profile inferred by iDRW (samples x pathways).
\item sigGenes : A list of the set of significant genes within pathways.
\item w : A final weight vector of genes by the product of directed random walk with restart on graphs.
\item pathSet : A list of the set of genes within pathways.
\item vertexZP : A data frame containing statistic score (\code{Score}) and p-value of the statistical test (\code{p-value}) for each gene.
}
}
\description{
This function infers pathway activities across samples using integrative directed random walk approach (iDRW) on multi-layered gene-gene graph.
}
\examples{
library(iDRW)
library(igraph)

g <- directGraph
c <- directGraph
m <- directGraph

gene_delim <- c('g.', 'c.', 'm.') # genes from RNA-Seq gene expression(g), CNV(c), Methylation(m) profile

V(g)$name <- paste(gene_delim[1],V(g)$name,sep="")
V(c)$name <-paste(gene_delim[2],V(c)$name,sep="")
V(m)$name <-paste(gene_delim[3],V(m)$name,sep="")

gcm <- (g \%du\% c) \%du\% m

class.outcome <- "time"
covs <- c("age", "gender", "stageT", "stageN", "stageM")
family <- "cox"

pa <- get.iDRWP(x=list(exp, cna, methyl), y=clinical, globalGraph=gcm, pathSet=pathSet, class.outcome=class.outcome,
                covs=covs, family=family, Gamma=0.3, Corr=FALSE)



}
